Further Turret  and shot stuff
Added collision detection, movement by time, and reading of arbitrary
turret image files

+++ AnimatedBar.java
@@ -7,7 +7,7 @@ public class AnimatedBar {
 	int size;
 	Color barColor;
 
-	public AnimatedBar(int pMax, int pSize, Color fillColor) {
+	public AnimatedBar(double pMax, int pSize, Color fillColor) {
 		max = pMax;
 		current = pMax;
 		size = pSize;
